Move paste functions to separate file due to their combined length.
