+++ src/kevedit/Makefile.am
@@ -17,8 +17,8 @@ endif
 EXTRA_DIST = macos.h
 
 kevedit_SOURCES = \
-	kevedit.c main.c menu.c misc.c patbuffer.c screen.c hash.c \
-	kevedit.h menu.h misc.h patbuffer.h screen.h hash.h $(windres_SOURCES) \
+	kevedit.c main.c menu.c misc.c patbuffer.c screen.c hash.c paste.c \
+	kevedit.h menu.h misc.h patbuffer.h screen.h hash.h paste.h $(windres_SOURCES) \
 	../mkdtemp/mkdtemp.c $(macos_SOURCES)
 
 kevedit_LDADD = ../zlaunch/libzlaunch.a ../texteditor/libtexteditor.a ../dialogs/libdialogs.a ../help/libhelp.a ../themes/rusty/librusty.a ../synth/libsynth.a ../display/libdisplay.a ../structures/libstructures.a ../libzzt2/libzzt2.a $(GLOBLIB)
